+++ app/src/main/java/com/app/arcabyolimpo/presentation/ui/components/atoms/inputs/SelectInput.kt
@@ -54,7 +54,7 @@ fun SelectInput(
         Text(
             text = label,
             style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurface,
+            color = White,
         )
 
         ExposedDropdownMenuBox(